requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorlover==0.2.1
cufflinks==0.14.6
cycler==0.10.0
dash==0.35.0
dash-core-components==0.41.0
dash-html-components==0.13.4
dash-renderer==0.16.1
dash-table==3.1.11
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonmerge==1.5.2
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.3.0
mysql-connector==2.1.6
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.4
numpy==1.15.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
plotly==3.4.2
pluggy==0.8.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.0
pytest==4.0.2
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
requests==2.21.0
retrying==1.3.3
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
